+++ azext_iot/__init__.py
@@ -26,12 +26,10 @@ class IoTExtCommandsLoader(AzCommandsLoader):
         from azext_iot.central.command_map import load_central_commands
         from azext_iot.digitaltwins.command_map import load_digitaltwins_commands
         from azext_iot.dps.command_map import load_dps_commands
-
-        if os.environ.get("IOT_CLI_ADU_ENABLED"):
-            from azext_iot.deviceupdate.command_map import load_deviceupdate_commands
-            load_deviceupdate_commands(self, args)
+        from azext_iot.deviceupdate.command_map import load_deviceupdate_commands
 
         load_command_table(self, args)
+        load_deviceupdate_commands(self, args)
         load_iothub_commands(self, args)
         load_central_commands(self, args)
         load_digitaltwins_commands(self, args)